[WIP][DsplayBatteryMode1] compile fail state

+++ firmware/atom_s3_battery_display/lib/display_battery_mode1/battery_display.h
@@ -6,7 +6,7 @@ class BatteryDisplay
 {
 public:
   BatteryDisplay(){};
-  BatteryDisplay(float bat_cell){bat_cell_ = bat_cell;}
+  BatteryDisplay(int bat_cell){bat_cell_ = bat_cell;}
   int getBatCell() {return bat_cell_;}
   void init(){};
